aws/requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiomultiprocess==0.9.0
aiosignal==1.2.0
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
bc-python-hcl2==0.3.45
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.24.52
botocore==1.27.52
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
checkov==2.1.137
click==8.0.4
click-config-file==0.6.0
click-option-group==0.5.3
cloudsplaining==0.5.0
colorama==0.4.5
ConfigArgParse==1.5.3
configobj==5.0.6
contextlib2==21.6.0
contourpy==1.0.5
cycler==0.11.0
cyclonedx-python-lib==2.7.1
debugpy==1.6.3
decorator==5.1.1
deep-merge==0.0.4
defusedxml==0.7.1
Deprecated==1.2.13
detect-secrets==1.3.0
docker==5.0.3
dockerfile-parse==1.2.0
dpath==1.5.0
entrypoints==0.4
executing==1.1.0
fastjsonschema==2.16.2
fonttools==4.37.3
frozenlist==1.3.1
gh-search==0.8.3
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpath-ng==1.5.3
jsonschema==3.2.0
junit-xml==1.9
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lark==1.1.2
lizard==1.17.10
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.4.12
numpy==1.23.3
packageurl-python==0.10.1
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
ply==3.11
policy-sentry==0.12.4
policyuniverse==1.5.0.20220613
prettytable==3.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycares==4.2.2
pycep-parser==0.3.8
pycparser==2.21
PyDriller==2.1
PyGithub==1.54.1
Pygments==2.13.0
PyJWT==1.7.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.7.25
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
schema==0.7.5
searchgithubrepo==1.0.10
semantic-version==2.10.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.5.1
tabulate==0.8.10
termcolor==1.1.0
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
tqdm==4.64.0
traitlets==5.4.0
types-pytz==2022.1.2
typing_extensions==4.3.0
update-checker==0.18.0
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
widgetsnbextension==4.0.3
wrapt==1.14.1
yarl==1.8.1
